feat: ajout typage des modeles mongo (#2790)

+++ server/src/common/actions/organismes/organismes.actions.ts
@@ -502,6 +502,7 @@ export const deleteOrganismeAndEffectifsAndDossiersApprenantsMigration = async (
 
   const organisme = await organismesDb().findOne({ _id });
   if (!organisme) throw new Error(`Unable to find organisme ${_id.toString()}`);
+  if (!organisme.uai) throw new Error(`Organisme ${_id.toString()} doesn't have any UAI`);
 
   // Suppression des dossiersApprenants liés sur la base du couple uai siret
   const { deletedCount: deletedDossiersApprenantsMigration } = await dossiersApprenantsMigrationDb().deleteMany({
